refactor(controller/source_folders): move browser model to tree.rs and localize fs/selection helpers
